Overhaul lend_cur()
Made much of the creation of orders more direct.
TODO: Requires more work on when it reaches the end of how many offers it requested.
Fixes #165 - This error was due to splitting a satoshi in half, and they both wanted to round up. Unwanted result of fix is an occasional satoshi or two left in balance. TODO: Tack on extra sats to last offer.